+++ CHANGES.md
@@ -1,6 +1,6 @@
-# Development version
+# 2022-08-09, 4.7.0
 
-- Support for Clang/LLVM 14.0.{1,2,3,4}
+- Support for Clang/LLVM 14.0.{1,2,3,4,5,6}
 
 - Support for `ParenList` expressions (may appear in constructor initializers
   of templated classes).
@@ -16,6 +16,10 @@
 
 - Fix lossing template specification in method definitions.
   (reported by Arthur Charguéraud)
+
+- Add `Clang.Cursor.{Set, Map}`, fix `Clang.compare_cursors`
+
+- Add `Clang.Lazy.Printer` module
 
 # 2022-03-22, 4.6.0